0.5.0 - Limited NMM Capabilities
You now can use this script with nexus mod manager.  Due to the nature
of texture overrides it will rely on the user to keep his data folder
secure.  I provide timestamps and backups of every transfer, so if
something goes wrong or a FaceGen file gets overwritten the user can
easily look up the folder to grab a spare.